refactor : splash, Authentication view, viewController 분리
splash, Authentication view, viewController 분리
