+++ NiceFilmTicket/Presentation/Authentication/SignIn/View/AuthenticationViewController.swift
@@ -10,92 +10,72 @@ import SnapKit
 
 class AuthenticationViewController: UIViewController {
     
-    
-    lazy var personalLoginImage = UIImageView()
-    lazy var businessLoginImage = UIImageView()
-    lazy var kakaoImageView = UIImageView()
-    lazy var idTextField = UITextField()
-    lazy var pwTextField = UITextField()
-    lazy var loginButton = UIButton()
-    lazy var orDivider = UIImageView()
-    lazy var signupButton = UIButton()
-    lazy var findId = UILabel()
-    lazy var findPw = UILabel()
-    lazy var dividerLine = UIImageView()
-    lazy var scrollView = UIScrollView()
-    lazy var logoImageView = UIImageView()
-    lazy var contentView = UIView()
-    lazy var personalLoginText = UILabel()
-    lazy var businessLoginText = UILabel()
+    private var authenticationView: AuthenticationView!
     
     override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .white
-        
-        configureLogoImageView()
-        configureScrollView(logoImageView: logoImageView)
-        setGeneralLogin()
-        // 개인 버튼 탭 제스처 인식기 생성 및 추가
-        let personalTextClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickPersonButton))
-        personalLoginText.addGestureRecognizer(personalTextClickRecognizer)
-        
-        let personalImageClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickPersonButton))
-        personalLoginImage.addGestureRecognizer(personalImageClickRecognizer)
-        
-        setBusinessLogin()
-        
-        // 비지니스 버튼 탭 제스처 인식기 생성 및 추가
-        let BusinessTextClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickBusinessButton))
-        businessLoginText.addGestureRecognizer(BusinessTextClickRecognizer)
-        
-        let BusinessImageClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickBusinessButton))
-        businessLoginImage.addGestureRecognizer(BusinessImageClickRecognizer)
-        
-        kakaoLoginButton()
-        iDTextField()
-        passwordTextField()
-        hideKeyboardWhenTappedAround()
-        createLoginButton()
-        createOrDivider()
-        createSignUpButton()
-        createDivider()
-        createFindIDButton()
-        createFindPWButton()
-    }
+            super.viewDidLoad()
+            
+            authenticationView = AuthenticationView(frame: view.bounds)
+            view.addSubview(authenticationView)
+            
+            setupGestureRecognizers()
+        }
 }
 
 extension AuthenticationViewController {
+    private func setupGestureRecognizers() {
+            // 개인 버튼 탭 제스처 인식기 생성 및 추가
+            let personalTextClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickPersonButton))
+            authenticationView.personalLoginText.addGestureRecognizer(personalTextClickRecognizer)
+            
+            let personalImageClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickPersonButton))
+            authenticationView.personalLoginImage.addGestureRecognizer(personalImageClickRecognizer)
+            
+             // 비지니스 버튼 탭 제스처 인식기 생성 및 추가
+             let businessTextClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickBusinessButton))
+             authenticationView.businessLoginText.addGestureRecognizer(businessTextClickRecognizer)
+
+             let businessImageClickRecognizer = UITapGestureRecognizer(target: self, action: #selector(clickBusinessButton))
+             authenticationView.businessLoginImage.addGestureRecognizer(businessImageClickRecognizer)
+
+             // 회원가입 버튼 이벤트 설정
+             authenticationView.signupButton.addTarget(self, action: #selector(didTapSignupButton), for:.touchUpInside)
+             
+             // 키보드 내리기 설정
+             hideKeyboardWhenTappedAround()
+        }
+    
     //개인 로그인으로 변경
     @objc private func clickPersonButton() {
-        personalLoginImage.image = UIImage(systemName: "person.fill")
-        businessLoginImage.image = UIImage(systemName: "bag")
-        kakaoImageView.isHidden = false
-        idTextField.isHidden = true
-        pwTextField.isHidden = true
-        loginButton.isHidden = true
-        orDivider.isHidden = true
-        signupButton.isHidden = true
-        findId.isHidden = true
-        findPw.isHidden = true
-        dividerLine.isHidden = true
-        scrollView.isScrollEnabled = false
+        authenticationView.personalLoginImage.image = UIImage(systemName: "person.fill")
+        authenticationView.businessLoginImage.image = UIImage(systemName: "bag")
+        authenticationView.kakaoImageView.isHidden = false
+        authenticationView.idTextField.isHidden = true
+        authenticationView.pwTextField.isHidden = true
+        authenticationView.loginButton.isHidden = true
+        authenticationView.orDivider.isHidden = true
+        authenticationView.signupButton.isHidden = true
+        authenticationView.findId.isHidden = true
+        authenticationView.findPw.isHidden = true
+        authenticationView.dividerLine.isHidden = true
+        authenticationView.scrollView.isScrollEnabled = false
     }
     
     //비즈니스 로그인으로 변경
     @objc private func clickBusinessButton() {
 
-        businessLoginImage.image = UIImage(systemName: "bag.fill")
-        personalLoginImage.image = UIImage(systemName: "person")
-        kakaoImageView.isHidden = true
-        idTextField.isHidden = false
-        pwTextField.isHidden = false
-        loginButton.isHidden = false
-        orDivider.isHidden = false
-        signupButton.isHidden = false
-        findId.isHidden = false
-        findPw.isHidden = false
-        dividerLine.isHidden = false
-        scrollView.isScrollEnabled = true
+        authenticationView.businessLoginImage.image = UIImage(systemName: "bag.fill")
+        authenticationView.personalLoginImage.image = UIImage(systemName: "person")
+        authenticationView.kakaoImageView.isHidden = true
+        authenticationView.idTextField.isHidden = false
+        authenticationView.pwTextField.isHidden = false
+        authenticationView.loginButton.isHidden = false
+        authenticationView.orDivider.isHidden = false
+        authenticationView.signupButton.isHidden = false
+        authenticationView.findId.isHidden = false
+        authenticationView.findPw.isHidden = false
+        authenticationView.dividerLine.isHidden = false
+        authenticationView.scrollView.isScrollEnabled = true
     }
     
     //회원가입 뷰로 이동
@@ -113,242 +93,5 @@ extension AuthenticationViewController {
     
     @objc func dismissKeyboard() {
         view.endEditing(true)
-    }
-    
-    //로고 만들기
-    func configureLogoImageView() {
-        logoImageView = UIImageView(image: UIImage(named: "Logo"))
-        view.addSubview(logoImageView)
-        
-        logoImageView.snp.makeConstraints { make in
-            make.centerY.equalTo(view.snp.top).offset(170)
-            make.centerX.equalTo(view.snp.centerX)
-            make.width.equalTo(210)
-            make.height.equalTo(150)
-        }
-    }
-    
-    // 스크롤뷰 설정
-    private func configureScrollView(logoImageView: UIImageView) {
-        scrollView = UIScrollView()
-        scrollView.backgroundColor = .white
-        scrollView.isScrollEnabled = false
-        scrollView.showsVerticalScrollIndicator = false
-        view.addSubview(scrollView)
-        
-        scrollView.snp.makeConstraints { make in
-            make.leading.equalTo(view.snp.leading)
-            make.trailing.equalTo(view.snp.trailing)
-            make.top.equalTo(logoImageView.snp.bottom) // logoImageView 아래에서 시작
-            make.bottom.equalTo(view.snp.bottom)
-            make.width.equalTo(view.snp.width)
-        }
-        
-        contentView = UIView()
-        scrollView.addSubview(contentView)
-        
-        contentView.snp.makeConstraints { make in
-            make.edges.equalTo(scrollView.contentLayoutGuide)
-            make.width.equalTo(scrollView.frameLayoutGuide)
-            make.height.equalTo(580)
-        }
-    }
-    
-    private func setGeneralLogin(){
-        // 개인로그인 이미지 뷰 생성
-        personalLoginImage = UIImageView(image: UIImage(systemName: "person.fill"))
-        personalLoginImage.isUserInteractionEnabled = true
-        personalLoginImage.tintColor = UIColor(red: 8/255, green: 30/255, blue: 92/255, alpha: 1)
-        contentView.addSubview(personalLoginImage)
-        
-        personalLoginImage.snp.makeConstraints { make in
-            make.width.equalTo(40)
-            make.height.equalTo(40)
-            make.leading.equalTo(contentView.snp.leading).offset(60)
-            make.top.equalTo(contentView.snp.top).offset(50)
-        }
-        
-        // 개인로그인 텍스트 레이블 생성
-        personalLoginText = MenuUILabel(text: "일반", size: UIFont.boldSystemFont(ofSize: 25))
-        personalLoginText.isUserInteractionEnabled = true
-        contentView.addSubview(personalLoginText)
-        
-        // 개인로그인 텍스트 제약 조건
-        personalLoginText.snp.makeConstraints { make in
-            make.leading.equalTo(personalLoginImage.snp.trailing).offset(0)
-            make.top.equalTo(contentView.snp.top).offset(55)
-        }
-    }
-    
-    private func setBusinessLogin() {
-        // 비지니스로그인 텍스트 레이블 생성
-        businessLoginText = MenuUILabel(text: "비즈니스", size: UIFont.boldSystemFont(ofSize: 25))
-        businessLoginText.isUserInteractionEnabled = true
-        contentView.addSubview(businessLoginText)
-        
-        // 비지니스로그인 텍스트 제약 조건
-        businessLoginText.snp.makeConstraints { make in
-            make.trailing.equalTo(contentView.snp.trailing).inset(60)
-            make.top.equalTo(contentView.snp.top).offset(55)
-        }
-        
-        // 비지니스로그인 이미지 뷰 생성
-        businessLoginImage = UIImageView(image: UIImage(systemName: "bag"))
-        businessLoginImage.isUserInteractionEnabled = true
-        businessLoginImage.tintColor = UIColor(red: 8/255, green: 30/255, blue: 92/255, alpha: 1)
-        contentView.addSubview(businessLoginImage)
-        
-        businessLoginImage.snp.makeConstraints { make in
-            make.trailing.equalTo(businessLoginText.snp.leading)
-            make.top.equalTo(contentView.snp.top).offset(50)
-            make.width.equalTo(40)
-            make.height.equalTo(40)
-        }
-    }
-    
-    private func kakaoLoginButton() {
-        //카카오 로그인 버튼
-        kakaoImageView = UIImageView(image: UIImage(named: "KakaoLogin"))
-        kakaoImageView.isUserInteractionEnabled = true
-        contentView.addSubview(kakaoImageView)
-        
-        // 카카오 로그인 뷰의 제약 조건 설정
-        kakaoImageView.snp.makeConstraints { make in
-            make.top.equalTo(personalLoginText.snp.bottom).offset(150)
-            make.centerX.equalTo(contentView.snp.centerX)
-            make.width.equalTo(250)
-            make.height.equalTo(60)
-        }
-    }
-    
-    private func iDTextField() {
-        //ID 입력 텍스트 필드
-        idTextField = AuthenticationTextField(placeholder: "  ID", isSecureTextEntry: false, backgroundColor: UIColor(red: 242/255, green: 242/255, blue: 242/255, alpha: 1), isHidden: true, font: UIFont.systemFont(ofSize: 20))
-        contentView.addSubview(idTextField)
-
-        idTextField.snp.makeConstraints { make in
-            make.height.equalTo(50)
-            make.top.equalTo(businessLoginImage.snp.bottom).offset(25)
-            make.leading.equalTo(contentView.snp.leading).inset(30)
-            make.trailing.equalTo(contentView.snp.trailing).inset(30)
-        }
-    }
-    
-    private func passwordTextField() {
-        //비밀번호 입력 텍스트 필드
-        pwTextField = AuthenticationTextField(placeholder: "  PW", isSecureTextEntry: true, backgroundColor: UIColor(red: 217/255, green: 217/255, blue: 217/255, alpha: 1), isHidden: true, font: UIFont.systemFont(ofSize: 20))
-        contentView.addSubview(pwTextField)
-
-        pwTextField.snp.makeConstraints { make in
-            make.top.equalTo(idTextField.snp.bottom).offset(20)
-            make.height.equalTo(50)
-            make.leading.equalTo(contentView.snp.leading).inset(30)
-            make.trailing.equalTo(contentView.snp.trailing).inset(30)
-        }
-    }
-    
-    private func createLoginButton() {
-        //로그인 버튼
-        loginButton = AuthenticationUIButton(title: "LogIn", isHidden: true)
-        contentView.addSubview(loginButton)
-
-        loginButton.snp.makeConstraints { make in
-            make.top.equalTo(pwTextField.snp.bottom).offset(20)
-            make.centerX.equalTo(contentView.snp.centerX)
-            make.width.equalTo(150)
-            make.height.equalTo(40)
-        }
-    }
-    
-    private func createOrDivider() {
-        //login버튼 아래 or
-        orDivider = UIImageView(image: UIImage(named: "OrDivider"))
-        orDivider.isHidden = true
-        contentView.addSubview(orDivider)
-
-        orDivider.snp.makeConstraints { make in
-            make.top.equalTo(loginButton.snp.bottom).offset(30)
-            make.centerX.equalTo(contentView.snp.centerX)
-            make.width.equalTo(280)
-            make.height.equalTo(10)
-        }
-    }
-    
-    private func createSignUpButton() {
-        //회원가입 버튼
-        signupButton = AuthenticationUIButton(title: "Make Account", isHidden: true)
-        signupButton.addTarget(self, action: #selector(didTapSignupButton), for: .touchUpInside) //회원가입 버튼 눌렀을 때
-        contentView.addSubview(signupButton)
-
-        signupButton.snp.makeConstraints { make in
-            make.top.equalTo(orDivider.snp.bottom).offset(30)
-            make.centerX.equalTo(contentView.snp.centerX)
-            make.width.equalTo(220)
-            make.height.equalTo(50)
-        }
-    }
-    
-    private func createDivider() {
-        //아이디 찾기 비밀번호 찾기 분리선
-        dividerLine = UIImageView(image: UIImage(named: "Line"))
-        dividerLine.isHidden = true
-        contentView.addSubview(dividerLine)
-
-        dividerLine.snp.makeConstraints { make in
-            make.top.equalTo(signupButton.snp.bottom).offset(85)
-            make.centerX.equalTo(contentView.snp.centerX)
-            make.width.equalTo(3)
-            make.height.equalTo(50)
-        }
-    }
-    
-    private func createFindIDButton() {
-        //아이디 찾기
-        findId = FindIdPwUILabel(text: "아이디 찾기")
-        contentView.addSubview(findId)
-
-        // 아이디 찾기 제약 조건
-        findId.snp.makeConstraints { make in
-            make.trailing.equalTo(dividerLine.snp.leading).offset(-65)
-            make.top.equalTo(signupButton.snp.bottom).offset(100)
-        }
-    }
-    
-    // 스크롤뷰 설정
-    private func configureScrollView(logoImgView: UIImageView) {
-        scrollView = UIScrollView()
-        scrollView.backgroundColor = .white
-        scrollView.isScrollEnabled = false
-        scrollView.showsVerticalScrollIndicator = false
-        view.addSubview(scrollView)
-        
-        scrollView.snp.makeConstraints { make in
-            make.leading.equalTo(view.snp.leading)
-            make.trailing.equalTo(view.snp.trailing)
-            make.top.equalTo(logoImgView.snp.bottom) // logoImageView 아래에서 시작
-            make.bottom.equalTo(view.snp.bottom)
-            make.width.equalTo(view.snp.width)
-        }
-        
-        contentView = UIView()
-        scrollView.addSubview(contentView)
-        
-        contentView.snp.makeConstraints { make in
-            make.edges.equalTo(scrollView.contentLayoutGuide)
-            make.width.equalTo(scrollView.frameLayoutGuide)
-            make.height.equalTo(580)
-        }
-    }
-    
-    private func createFindPWButton() {
-        //비밀번호 찾기
-        findPw = FindIdPwUILabel(text: "비밀번호 찾기")
-        contentView.addSubview(findPw)
-
-        // 비밀번호 찾기 제약 조건
-        findPw.snp.makeConstraints { make in
-            make.leading.equalTo(dividerLine.snp.trailing).offset(60)
-            make.top.equalTo(signupButton.snp.bottom).offset(100)
-        }
     }
 }
